PY/Nonogram/requirements.txt
numpy==1.19.5
pygame==2.0.1
Pillow==8.3.2